+++ src/logan/pickpocket/commands/AdminCommand.java
@@ -12,11 +12,18 @@ import java.util.List;
  * Created by Tre on 1/10/2016.
  */
 public class AdminCommand implements PickpocketCommand {
+
+    private Pickpocket pickpocket;
+
+    public AdminCommand(Pickpocket pickpocket) {
+        this.pickpocket = pickpocket;
+    }
+
     @Override
     public void execute(Player player, List<Profile> profiles, Object... args) {
         if (player.isOp() || player.hasPermission(Pickpocket.PICKPOCKET_ADMIN)) {
             boolean bool = Boolean.parseBoolean(args[0].toString());
-            Profiles.get(player, profiles).getPermissionModule().setAdmin(bool);
+            Profiles.get(player, profiles,pickpocket).getProfileConfiguration().setAdminSection(bool);
             player.sendMessage(ChatColor.GRAY + "Pickpocket Admin status set to " + bool + ".");
         }
     }