+++ src/logan/pickpocket/commands/ItemsCommand.java
@@ -1,5 +1,6 @@
 package logan.pickpocket.commands;
 
+import logan.pickpocket.main.Pickpocket;
 import logan.pickpocket.main.Profiles;
 import logan.pickpocket.profile.Profile;
 import org.bukkit.entity.Player;
@@ -10,9 +11,16 @@ import java.util.List;
  * Created by Tre on 12/18/2015.
  */
 public class ItemsCommand implements PickpocketCommand {
+
+    private Pickpocket pickpocket;
+
+    public ItemsCommand(Pickpocket pickpocket) {
+        this.pickpocket = pickpocket;
+    }
+
     @Override
     public void execute(Player player, List<Profile> profiles, Object... objects) {
-        Profile profile = Profiles.get(player, profiles);
+        Profile profile = Profiles.get(player, profiles,pickpocket);
         profile.openPickpocketItemInventory();
     }
 }