+++ src/logan/pickpocket/main/Profiles.java
@@ -9,12 +9,16 @@ import java.util.List;
  * Created by Tre on 12/17/2015.
  */
 public abstract class Profiles {
-    public static Profile get(Player player, List<Profile> profiles) {
+    public static Profile get(Player player, List<Profile> profiles, Pickpocket pickpocket) {
         for (Profile profile : profiles) {
-            if (profile.getPlayer().equals(player)) {
+            if (profile.getPlayer().getUniqueId().equals(player.getUniqueId())) {
+                profile.setPlayer(player);
                 return profile;
             }
         }
-        return null;
+
+        Profile profile=new Profile(player,pickpocket);
+        Pickpocket.addProfile(profile);
+        return profile;
     }
 }