+++ src/logan/pickpocket/profile/PickpocketItemInventory.java
@@ -1,10 +1,14 @@
 package logan.pickpocket.profile;
 
+import logan.pickpocket.main.Pickpocket;
 import logan.pickpocket.main.PickpocketItem;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -14,9 +18,12 @@ import java.util.*;
 /**
  * Created by Tre on 12/26/2015.
  */
-public class PickpocketItemInventory {
+public class PickpocketItemInventory implements Listener {
 
     public static final String NAME = "Pickpocket Items";
+
+    private final String nextButtonName = ChatColor.GRAY + "Next";
+    private final String backButtonName = ChatColor.GRAY + "Back";
 
     private ItemStack nextButton;
     private ItemStack backButton;
@@ -27,8 +34,9 @@ public class PickpocketItemInventory {
     private Profile profile;
     private Player player;
 
-    public PickpocketItemInventory(Profile profile) {
+    public PickpocketItemInventory(Profile profile, Pickpocket pickpocket) {
         this.profile = profile;
+        pickpocket.getServer().getPluginManager().registerEvents(this, pickpocket);
     }
 
     public void open() {
@@ -40,7 +48,8 @@ public class PickpocketItemInventory {
 
         configureButtons();
 
-        if (!pickpocketItems.isEmpty()) inventoryList.add(Bukkit.createInventory(null, 54, "Pickpocket Items Page: 1"));
+        if (!pickpocketItems.isEmpty())
+            inventoryList.add(Bukkit.createInventory(null, 54, "Pickpocket Items Page: 1"));
 
         int slotNum = 0;
         int maxSize = 44;
@@ -81,14 +90,6 @@ public class PickpocketItemInventory {
         player.openInventory(addButtons(inventoryList.get(inventoryOpen)));
     }
 
-    public String getNextButtonName() {
-        return nextButton.getItemMeta().getDisplayName();
-    }
-
-    public String getBackButtonName() {
-        return backButton.getItemMeta().getDisplayName();
-    }
-
     public void nextPage() {
         if (inventoryOpen == inventoryList.size() - 1) return;
         player.openInventory(addButtons(inventoryList.get(++inventoryOpen)));
@@ -104,11 +105,11 @@ public class PickpocketItemInventory {
         backButton = new ItemStack(Material.PAPER);
 
         ItemMeta nextButtonMeta = nextButton.getItemMeta();
-        nextButtonMeta.setDisplayName(ChatColor.GRAY + "Next");
+        nextButtonMeta.setDisplayName(nextButtonName);
         nextButton.setItemMeta(nextButtonMeta);
 
         ItemMeta backButtonMeta = backButton.getItemMeta();
-        backButtonMeta.setDisplayName(ChatColor.GRAY + "Back");
+        backButtonMeta.setDisplayName(backButtonName);
         backButton.setItemMeta(backButtonMeta);
     }
 
@@ -116,6 +117,24 @@ public class PickpocketItemInventory {
         inventory.setItem(45, backButton);
         inventory.setItem(53, nextButton);
         return inventory;
+    }
+
+    @EventHandler
+    public void onInventoryClick(InventoryClickEvent e) {
+        try {
+            if (e.getInventory().getItem(e.getSlot()) == null) return;
+            if (e.getClickedInventory() == null) return;
+        } catch (Exception e1) {
+            return;
+        }
+        if (e.getCurrentItem().getItemMeta().getDisplayName().equals(nextButtonName)) {
+            nextPage();
+            e.setCancelled(true);
+        }
+        if (e.getCurrentItem().getItemMeta().getDisplayName().equals(backButtonName)) {
+            previousPage();
+            e.setCancelled(true);
+        }
     }
 
 }